refactor BannerText Interfaces
